// src/types/enums.ts
export enum ProvinceType {
  Land = "land",
  Sea = "sea",
  Lake = "lake",
  Wasteland = "wasteland",
}

export enum TerrainType {
  Plains = "plains",
  Forest = "forest",
  Hills = "hills",
  Mountain = "mountain",
  Desert = "desert",
  Marsh = "marsh",
  Jungle = "jungle",
  Urban = "urban",
  Ocean = "ocean",
  Coastal = "coastal",
  Arctic = "arctic",
  Savanna = "savanna",
  Steppe = "steppe",
  Farmland = "farmland",
  Tundra = "tundra",
  DeepOcean = "deep_ocean",
}

export enum ResourceType {
  Steel = "steel",
  Oil = "oil",
  Aluminium = "aluminium",
  Rubber = "rubber",
  Tungsten = "tungsten",
  Chromium = "chromium",
  Gold = "gold",
  Coal = "coal",
  Uranium = "uranium",
  Iron = "iron",
  Copper = "copper",
  Timber = "timber",
  Grain = "grain",
  Fish = "fish",
}

export enum StateCategory {
  Wasteland = "wasteland",
  Enclave = "enclave",
  TinyIsland = "tiny_island",
  Pastoral = "pastoral",
  Rural = "rural",
  Town = "town",
  LargeTown = "large_town",
  City = "city",
  LargeCity = "large_city",
  Metropolis = "metropolis",
  Megalopolis = "megalopolis",
  Megacity = "megacity",
}

export enum GovernmentType {
  Democracy = "democracy",
  Fascism = "fascism",
  Communism = "communism",
  Neutrality = "neutrality",
  Monarchy = "monarchy",
  Theocracy = "theocracy",
  Oligarchy = "oligarchy",
  MilitaryJunta = "military_junta",
  AbsoluteMonarchy = "absolute_monarchy",
  ConstitutionalMonarchy = "constitutional_monarchy",
}

export enum IdeologyType {
  Democratic = "democratic",
  Fascist = "fascist",
  Communist = "communist",
  Neutral = "neutral",
  Monarchist = "monarchist",
  Theocratic = "theocratic",
}

export enum AdjacencyType {
  Land = "land",
  Sea = "sea",
  River = "river",
  Strait = "strait",
  Canal = "canal",
  Impassable = "impassable",
}

export enum ToolType {
  Select = "select",
  Paint = "paint",
  Erase = "erase",
  Fill = "fill",
  Eyedropper = "eyedropper",
  RectangleSelect = "rectangle_select",
  TerrainPaint = "terrain_paint",
  StatePaint = "state_paint",
  CountryPaint = "country_paint",
}

export enum ViewMode {
  Province = "province",
  Political = "political",
  Terrain = "terrain",
  State = "state",
  Population = "population",
  Development = "development",
  Resource = "resource",
}

export enum BrushShape {
  Circle = "circle",
  Square = "square",
  Diamond = "diamond",
}